05 - Render Objects present in the array (not working solution)
This doesn't work on the first click because state update does not
happen immediately.
When you call getNextPerson function, it updates the currentPerson
state, but the new state value is not immediately available in the
following lines of code that use it to update other states. The
state updates are batched by React and applied asynchronously, so
the person and btnName states still use the old value of
currentPerson in the first click. On the second click, the new
currentPerson value is already available, so the states are updated
correctly.