10 - Project structure for large projects (Folder example)
In this structure the components are put into their own folder with
their css files. Inside the folder for component we have a index.jsx
file which is used to import everything related to one component and
export it.
index.jsx helps us to avoid component_name/compoent_name.jsx type of
path while importing the component into another file.